Fix scaling in block projection
Note: scaling should obviously be applied to coords before applying the SE3
transform. When aggregating multiple points in a single matrix, scaling
should be done with an homogenous 4x4 diagonal matrix in which each
entry is the scaling factor except the last one which should be 1,
otherwise the transform will break.