Fix handling of neighbors outside the volume
The Octree::get_face_neighbors() is now templated with a safe boolean
parameter. If safe is true, the case of neighbors outside the volume is
tested and empty is returned for them, sacrificing performance. If safe
is false, then the previous behavior of returning garbage is retained,
but performance is not reduced.